refactor(ui): rewrite repositories dropdown
