+++ ui/src/components/dropdowns/RepositoriesDropdown.tsx
@@ -1,25 +1,71 @@
-import React, { useState } from "react";
+import React, { useState, useRef } from "react";
+import { twMerge } from "tailwind-merge";
+import {
+  useFloating,
+  useClick,
+  useDismiss,
+  useRole,
+  useListNavigation,
+  useInteractions,
+  FloatingFocusManager,
+  offset,
+  autoUpdate,
+  FloatingPortal,
+} from "@floating-ui/react";
 import { useQuery } from "@tanstack/react-query";
 
 import { fetchRepositories } from "@/clients/repositories/client";
 import { reactQueryKeys } from "@/clients/reactQueryConfig";
 
-import Box from "@/components/core/Box";
+import Dropdown from "@/components/core/Dropdown";
 import Input from "@/components/core/Input";
 import Checkbox from "@/components/core/Checkbox";
 
 export interface RepositoriesDropdownProps {
+  className?: string;
   variant?: "light" | "dark";
-  filter: string[];
-  onChange: (filter: string[]) => void;
+  disabled?: boolean;
+  selectedRepositories: string[];
+  setSelectedRepositories: (repositories: string[]) => void;
 }
 
 const RepositoriesDropdown: React.FC<RepositoriesDropdownProps> = ({
-  variant,
-  filter,
-  onChange,
+  className,
+  variant = "light",
+  disabled,
+  selectedRepositories,
+  setSelectedRepositories,
 }) => {
+  const [isOpen, setIsOpen] = useState(false);
   const [search, setSearch] = useState<string>("");
+  const [activeIndex, setActiveIndex] = useState<number | null>(null);
+
+  const listElementsRef = useRef<Array<HTMLElement | null>>([]);
+
+  const { refs, floatingStyles, context } = useFloating<HTMLElement>({
+    placement: "bottom-start",
+    open: isOpen,
+    onOpenChange: setIsOpen,
+    whileElementsMounted: autoUpdate,
+    middleware: [offset(8)],
+  });
+
+  const click = useClick(context, {
+    enabled: !disabled,
+    event: "mousedown",
+  });
+  const listNavigation = useListNavigation(context, {
+    enabled: !disabled,
+    listRef: listElementsRef,
+    activeIndex: activeIndex,
+    onNavigate: setActiveIndex,
+  });
+  const dismiss = useDismiss(context);
+  const role = useRole(context, { role: "combobox" });
+
+  const { getReferenceProps, getFloatingProps, getItemProps } = useInteractions(
+    [click, listNavigation, dismiss, role]
+  );
 
   const repositoriesQuery = useQuery({
     queryKey: reactQueryKeys.repositories,
@@ -32,87 +78,134 @@ const RepositoriesDropdown: React.FC<RepositoriesDropdownProps> = ({
     }),
   });
 
-  const handleChange = (
-    e: React.ChangeEvent<HTMLInputElement>,
-    repository: string
-  ) => {
-    if (e.target.checked) {
-      onChange([...filter, repository]);
+  const handleSearch = (e: React.ChangeEvent<HTMLInputElement>) => {
+    setSearch(e.target.value);
+    setActiveIndex(null);
+  };
+
+  const handleSelect = (repository: string) => {
+    if (selectedRepositories.includes(repository)) {
+      setSelectedRepositories(
+        selectedRepositories.filter((r) => r !== repository)
+      );
     } else {
-      onChange(filter.filter((r) => r !== repository));
+      setSelectedRepositories([...selectedRepositories, repository]);
     }
   };
 
-  const handleSearch = (e: React.ChangeEvent<HTMLInputElement>) => {
-    setSearch(e.target.value);
+  const styles = {
+    light: `bg-nuances-white
+      text-primary-600
+      shadow-light`,
+    dark: `bg-nuances-black
+      text-nuances-300
+      shadow-dark`,
   };
 
   return (
-    <Box
-      variant={variant}
-      className={`
-        flex-col
-        items-center
-        justify-center
-        z-10
-        gap-2
-        ${variant === "light" ? "text-primary-600" : "text-nuances-300"}
-      `}
-    >
-      <span className="self-start mx-4 mt-2">Repositories</span>
-      <hr
-        className={`
-          h-[1px]
-          w-full
-          ${variant === "light" ? "border-primary-600" : "border-nuances-300"}
-        `}
-      />
-      <Input
+    <>
+      <Dropdown
+        className={className}
+        label="Repositories"
+        filled={selectedRepositories.length > 0}
+        disabled={disabled}
         variant={variant}
-        className="w-full mx-2"
-        placeholder="Search repositories"
-        value={search}
-        onChange={handleSearch}
+        forwardRef={refs.setReference}
+        {...getReferenceProps()}
       />
-      <hr
-        className={`
-          h-[1px]
-          w-full
-          ${variant === "light" ? "border-primary-600" : "border-nuances-300"}
-        `}
-      />
-      <div
-        className={`
-          flex
-          flex-col
-          self-start
-          max-h-52
-          w-full
-          overflow-scroll
-          px-4
-          py-1
-          mb-2
-          gap-2
-        `}
-      >
-        {repositoriesQuery.isLoading && <span>Loading...</span>}
-        {repositoriesQuery.isError && <span>An error occurred.</span>}
-        {repositoriesQuery.isSuccess &&
-          (repositoriesQuery.data.results.length !== 0 ? (
-            repositoriesQuery.data.results.map((repository) => (
-              <Checkbox
-                key={repository.name}
-                variant={variant}
-                label={repository.name}
-                checked={filter.includes(repository.name)}
-                onChange={(e) => handleChange(e, repository.name)}
+      {isOpen && (
+        <FloatingPortal>
+          <FloatingFocusManager
+            context={context}
+            modal={false}
+            initialFocus={-1}
+          >
+            <div
+              ref={refs.setFloating}
+              style={floatingStyles}
+              className={twMerge(
+                `overflow-y-auto
+                rounded-lg
+                outline-none
+                px-4
+                py-2`,
+                styles[variant]
+              )}
+              {...getFloatingProps()}
+            >
+              <span className="font-semibold">Repositories</span>
+              <hr
+                className={`
+                  h-[1px]
+                  w-auto
+                  -mx-4
+                  my-2
+                  ${
+                    variant === "light"
+                      ? "border-primary-600"
+                      : "border-nuances-300"
+                  }
+                `}
               />
-            ))
-          ) : (
-            <span>No repositories found.</span>
-          ))}
-      </div>
-    </Box>
+              <Input
+                variant={variant}
+                className="-mx-2"
+                placeholder="Search repositories"
+                value={search}
+                onChange={handleSearch}
+              />
+              <hr
+                className={`
+                  h-[1px]
+                  w-auto
+                  -mx-4
+                  my-2
+                  ${
+                    variant === "light"
+                      ? "border-primary-600"
+                      : "border-nuances-300"
+                  }
+                `}
+              />
+              <div className="flex flex-col gap-1">
+                {repositoriesQuery.isLoading && <span>Loading...</span>}
+                {repositoriesQuery.isError && <span>An error occurred.</span>}
+                {repositoriesQuery.isSuccess &&
+                  (repositoriesQuery.data.results.length !== 0 ? (
+                    repositoriesQuery.data.results.map((repository, index) => (
+                      <Checkbox
+                        key={repository.name}
+                        role="option"
+                        variant={variant}
+                        label={repository.name}
+                        checked={selectedRepositories.includes(repository.name)}
+                        readOnly
+                        tabIndex={activeIndex === index ? 0 : -1}
+                        forwardedRef={(node) => {
+                          listElementsRef.current[index] = node;
+                        }}
+                        {...getItemProps({
+                          onClick() {
+                            handleSelect(repository.name);
+                          },
+                          onKeyDown(event) {
+                            if (event.key === "Enter" || event.key === " ") {
+                              event.preventDefault();
+                              handleSelect(repository.name);
+                            }
+                          },
+                        })}
+                      />
+                    ))
+                  ) : (
+                    <span>No repositories found.</span>
+                  ))}
+              </div>
+            </div>
+          </FloatingFocusManager>
+        </FloatingPortal>
+      )}
+    </>
   );
 };
